Make exp type more consistent

+++ include/boost/decimal/decimal128.hpp
@@ -155,7 +155,7 @@ BOOST_DECIMAL_EXPORT class decimal128 final
 {
 public:
     using significand_type = detail::uint128;
-    using exponent_type = std::uint64_t;
+    using exponent_type = std::uint32_t;
     using biased_exponent_type = std::int32_t;
 
 private:
